+++ app/src/main/java/org/welbodipartnership/cradle5/facilities/otherinfoform/FacilityOtherInfoFormViewModel.kt
@@ -33,7 +33,7 @@ class FacilityOtherInfoFormViewModel @Inject constructor(
     class Error(val errorMessage: String) : FormState()
   }
 
-  private val existingFacilityPrimaryKey: Long = requireNotNull(
+  private val existingFacilityPrimaryKey: Int = requireNotNull(
     handle[LeafScreen.FacilityDetails.ARG_FACILITY_PRIMARY_KEY]
   ) { "missing existingFacilityPrimaryKey" }
